+++ API/src/main/java/fr/maxlego08/essentials/api/hologram/AutoUpdateTaskConfiguration.java
@@ -1,0 +1,7 @@
+package fr.maxlego08.essentials.api.hologram;
+
+public record AutoUpdateTaskConfiguration (
+        boolean enable,
+        long milliseconds
+){
+}