+++ src/main/java/fr/maxlego08/essentials/user/ZUser.java
@@ -1,6 +1,6 @@
 package fr.maxlego08.essentials.user;
 
-import com.tcoded.folialib.impl.ServerImplementation;
+import com.tcoded.folialib.impl.PlatformScheduler;
 import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.Permission;
 import fr.maxlego08.essentials.api.dto.CooldownDTO;
@@ -87,6 +87,7 @@ public class ZUser extends ZUtils implements User {
     private long offlineVote;
     private Map<String, Long> lastVotes = new HashMap<>();
     private Home currentDeleteHome;
+    private long flySeconds;
 
     private boolean freeze;
 
@@ -230,13 +231,13 @@ public class ZUser extends ZUtils implements User {
         Location playerLocation = getPlayer().getLocation();
         AtomicInteger atomicInteger = new AtomicInteger(teleportationModule.getTeleportationDelay(getPlayer()));
 
-        if (teleportationModule.isTeleportDelayBypass() && this.hasPermission(Permission.ESSENTIALS_TELEPORT_BYPASS)) {
+        if (teleportationModule.isTeleportDelayBypass() && this.hasPermission(Permission.ESSENTIALS_TELEPORT_BYPASS) || atomicInteger.get() <= 0) {
             this.teleport(teleportationModule, location, successMessage, args);
             return;
         }
 
-        ServerImplementation serverImplementation = this.plugin.getScheduler();
-        serverImplementation.runAtLocationTimer(location, wrappedTask -> {
+        PlatformScheduler platformScheduler = this.plugin.getScheduler();
+        platformScheduler.runAtLocationTimer(location, wrappedTask -> {
 
             if (!this.isOnline()) {
                 wrappedTask.cancel();
@@ -771,6 +772,7 @@ public class ZUser extends ZUtils implements User {
         this.playTime = userDTO.play_time();
         this.lastLocation = stringAsLocation(userDTO.last_location());
         this.freeze = userDTO.frozen() != null && userDTO.frozen();
+        this.flySeconds = userDTO.fly_seconds();
     }
 
     @Override
@@ -850,12 +852,35 @@ public class ZUser extends ZUtils implements User {
     }
 
     @Override
+    public Optional<Home> getCurrentDeleteHome() {
+        return Optional.ofNullable(this.currentDeleteHome);
+    }
+
+    @Override
     public void setCurrentDeleteHome(Home currentDeleteHome) {
         this.currentDeleteHome = currentDeleteHome;
     }
 
     @Override
-    public Optional<Home> getCurrentDeleteHome() {
-        return Optional.ofNullable(this.currentDeleteHome);
+    public long getFlySeconds() {
+        return this.flySeconds;
+    }
+
+    @Override
+    public void setFlySeconds(long seconds) {
+        this.flySeconds = seconds;
+        getStorage().upsertFlySeconds(this.uniqueId, this.flySeconds);
+    }
+
+    @Override
+    public void addFlySeconds(long seconds) {
+        this.flySeconds += seconds;
+        getStorage().upsertFlySeconds(this.uniqueId, this.flySeconds);
+    }
+
+    @Override
+    public void removeFlySeconds(long seconds) {
+        this.flySeconds -= seconds;
+        getStorage().upsertFlySeconds(this.uniqueId, this.flySeconds);
     }
 }